+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Features and Fixes
 
+- fix: Crash when passing garbage to maxBreadcrumbs #1018
 - fix: OutOfMemory exception type #1015
 - fix: macOS version for Mac Catalyst #1011